requirements.txt
openai
pdfminer
requests
python-telegram-bot
whisper
python-docx
python-pptx
